CRUD

+++ src/main/java/com/rms/startup/Service/CategoryService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.rms.startup.Messages;
 import com.rms.startup.Bean.CategoryBean;
 import com.rms.startup.DAO.CategoryDAO;
 
@@ -18,21 +19,40 @@ public class CategoryService {
 		return CategoryDAO.getAllCategory();
 	}
 
-	public void addCategory(CategoryBean bean) {
-		CategoryDAO.addCategory(bean);
+	public String addCategory(CategoryBean bean) {
+		if(CategoryDAO.getCategory(bean.getCategoryId()) == null)
+		{
+			CategoryDAO.addCategory(bean);
+			return Messages.added;
+		}
+		return Messages.alreadyExist;
 	}
 
-	public void deleteCategory(Integer categoryId) {
-		CategoryDAO.deleteCategory(categoryId);
+	public String deleteCategory(Integer categoryId) {
+		if(CategoryDAO.getCategory(categoryId) != null)
+		{
+				try {
+			CategoryDAO.deleteCategory(categoryId);
+			return Messages.deleted;
+			} catch (Exception e) {
+				return Messages.itemExistError; 
+			}
+		}
+		else
+			return Messages.doesNotExist;
 	}
 
 	public CategoryBean getCategory(Integer categoryId) {
-		System.out.println("CategoryId Service " + categoryId);
 		return CategoryDAO.getCategory(categoryId);
 	}
 
-	public void updateCategory(Integer categoryId, CategoryBean bean) {
-		CategoryDAO.updateCategory(categoryId, bean);
+	public String updateCategory(CategoryBean bean) {
+		if(CategoryDAO.getCategory(bean.getCategoryId()) != null)
+		{
+			CategoryDAO.updateCategory(bean);
+			return Messages.updated;
+		}
+		return Messages.doesNotExist;
 	}
 
 }
